requirements.txt
PyQt6
keyboard
rapidocr_onnxruntime
thefuzz
python-Levenshtein
mss
pygetwindow
Pillow